popup fixed

+++ src/Componants/Popup.js
@@ -1,22 +1,41 @@
 import React, { useState } from 'react';
+import Swal from 'sweetalert2';
 
 function Popup({ closePopUp, selected, addToWish }) {
     const [wL, setWl] = useState(true)
 
     var remVal = selected.Title + "," + selected.Poster + "," + selected.imdbID
-    let lS = localStorage.getItem("myData")
-    // let accesse = lS.toString().search(remVal) > -1
+    let lS = localStorage.getItem("myData")//.replace(/\s+/g, "")
     const removeFromWish = () => {
-        let v = lS.replace(remVal, "").replace(/s+/g, "")
-        localStorage.setItem("myData", v)
-        setWl(true)
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+          }).then((result) => {
+            if (result.isConfirmed) {
+                let v = lS.replace(remVal, "")
+                localStorage.setItem("myData", v)
+                setWl(true)
+              Swal.fire(
+                'Deleted!',
+                'Your file has been deleted.',
+                'success'
+              )
+            }
+          })
     }
     return (
         <section className="popup">
             <div className="button-container" >
                 <button className="close" onClick={closePopUp}>close</button>
-                {wL ? <button className="close" onClick={() => { addToWish(selected); setWl(false) }}>Add To Wish List</button> :
-                    <button className="close" onClick={() => { removeFromWish(); }}>Remove From Wish List</button>}
+                <button className="close" onClick={() => { addToWish(selected); setWl(false) }}>Add To Wish List</button> 
+                    <button className="close" onClick={() => { removeFromWish(); }}>Remove From Wish List</button>
+                {/* {wL ? <button className="close" onClick={() => { addToWish(selected); setWl(false) }}>Add To Wish List</button> :
+                    <button className="close" onClick={() => { removeFromWish(); }}>Remove From Wish List</button>} */}
             </div>
             <div className="content">
                 <img src={selected.Poster} ></img>
